requirements.txt
moz-sql-parser
sqlparse
klepto
psycopg2-binary
numpy==1.16
pandas
torch
torchvision
tensorflow
networkx
matplotlib
seaborn
sklearn
toml
pyzmq
psutil
wget
xgboost
sqlalchemy
pytest
progressbar2
pomegranate
nltk
pygtrie
